requirements.txt
# ── Core LLM & Agent Orchestration ──────────────────────────────
groq
langgraph
langchain-core
langchain-community
langchain-huggingface
# ── Embeddings ───────────────────────────────────────────────────
sentence-transformers
# ── Vector Store ─────────────────────────────────────────────────
faiss-cpu
# ── Math Engine ──────────────────────────────────────────────────
sympy

# ── UI ───────────────────────────────────────────────────────────
streamlit

# ── Audio Processing ─────────────────────────────────────────────
# Groq Whisper is used via API — no local audio model needed
# pydub is used only for audio format conversion before sending to Groq
pydub

# ── Image Processing ─────────────────────────────────────────────
# Groq Vision (LLaVA) is used via API — no local OCR model needed
Pillow

# ── Environment & Config ─────────────────────────────────────────
python-dotenv
# ── Data Handling ────────────────────────────────────────────────
numpy
pandas